refactor: rename data to constants directory
