+++ src/api/spaceX.instance.ts
@@ -1,6 +1,6 @@
 import axios_ from 'axios';
 
-import { apiUrl } from '../data/apiUrl';
+import { apiUrl } from '../constants/apiUrl';
 
 export const instance = axios_.create({
   baseURL: apiUrl,
