+++ src/constants/texts.ts
