+++ src/store/store.ts
@@ -4,9 +4,9 @@ import produce, { Draft } from 'immer';
 import { Action, createHook, createStore, defaults } from 'react-sweet-state';
 
 import { ErrorRequestResponse, SpaceXTypes } from '../api/spaceX.types';
-import { cardsData } from '../data/cards';
-import { CardsID } from '../data/cards.types';
-import { errorMessages } from '../data/errorMessages';
+import { cardsData } from '../constants/cards';
+import { CardsID } from '../constants/cards.types';
+import { errorMessages } from '../constants/errorMessages';
 import { isDefined } from '../utils/utils';
 import { State } from './store.types';
